Remove redundant AbstractTestAlluxioFileSystem
